Compatability has been updated for more complex operations.
The "processButton" method now checks if the result was displayed before setting prevOperator. If resultDisplayed is true, it prepares for a new operation with the result without clearing prevOperator. Otherwise, it evaluates the previous operation before setting the new operator.

The "doMath4F" method now checks if the operator was pressed or if the result was displayed before updating the prevNumber. If not, it performs the previous operation with prevNumber and currentNumber.